Lakes SMRF gold ouputs are exactly what SMRF tests output, removed to just keep snow and em files
