+++ tests/basins/Lakes/config.ini
@@ -1,17 +1,4 @@
 ################################################################################
-# Config File for SMRF v0.5.3-42-g6b37dac
-# For more SMRF related help see:
-# http://smrf.readthedocs.io/en/latest/
-#
-# Date generated: 2018-07-13
-#
-# Generated using: inicheck <filename> -w
-#
-# For more inicheck help see:
-# http://inicheck.readthedocs.io/en/latest/
-
-
-################################################################################
 # Files for DEM and vegetation
 ################################################################################
 
@@ -125,3 +112,71 @@ input_backup:                  True
 threading:                     False
 log_level:                     debug
 log_file:                      ./output/log.txt
+
+################################################################################
+# Configurations for AWSM Master section
+################################################################################
+
+[awsm master]
+run_smrf:                      True
+model_type:                    ipysnobal
+
+
+################################################################################
+# Configurations for PATHS section for rigid directory work
+################################################################################
+
+[paths]
+path_dr:                       ./output
+basin:                         lakes
+isops:                         False
+proj:                          lakes_gold
+desc:                          Lakes gold HRRR simulation
+folder_date_style:             wyhr
+
+
+################################################################################
+# Configurations for GRID data to run iSnobal
+################################################################################
+
+[grid]
+
+
+################################################################################
+# Input files to run AWSM
+################################################################################
+
+[files]
+init_file:                     ./topo/snow.nc
+init_type:                     netcdf_out
+
+
+################################################################################
+# System parameters
+################################################################################
+
+[awsm system]
+log_level:                     debug
+log_to_file:                   True
+output_frequency:              1
+variables:                     thickness, snow_density, specific_mass, liquid_water, temp_surf, temp_lower, temp_snowcover, thickness_lower, water_saturation, net_rad, sensible_heat, latent_heat, snow_soil, precip_advected, sum_eb, evaporation, snowmelt, swi, cold_content
+
+
+################################################################################
+# Parameters for restarting from crash
+################################################################################
+
+[isnobal restart]
+
+################################################################################
+# Running Python wrapped iSnobal
+################################################################################
+
+[ipysnobal]
+
+################################################################################
+# Input constants for PySnobal
+################################################################################
+
+[ipysnobal constants]
+
