+++ tests/basins/Lakes/gold_hrrr/gold_config.ini
@@ -233,3 +233,91 @@ log_level:                     debug
 log_file:                      ./log.txt
 qotw:                          False
 queue_max_values:              2
+
+################################################################################
+# Configurations for AWSM Master section
+################################################################################
+
+[awsm master]
+run_smrf:                      True
+model_type:                    ipysnobal
+mask_isnobal:                  False
+prompt_dirs:                   False
+
+
+################################################################################
+# Configurations for PATHS section for rigid directory work
+################################################################################
+
+[paths]
+path_dr:                       ./
+basin:                         lakes
+isops:                         False
+proj:                          lakes_gold
+desc:                          Lakes gold HRRR simulation
+folder_date_style:             wyhr
+
+
+################################################################################
+# Configurations for GRID data to run iSnobal
+################################################################################
+
+[grid]
+active_layer:                  0.25
+thresh_normal:                 60
+thresh_medium:                 10
+thresh_small:                  1
+
+
+################################################################################
+# Input files to run AWSM
+################################################################################
+
+[files]
+init_file:                     ../topo/snow.nc
+init_type:                     netcdf_out
+
+
+################################################################################
+# System parameters
+################################################################################
+
+[awsm system]
+log_level:                     debug
+log_to_file:                   True
+run_for_nsteps:                None
+output_frequency:              1
+ithreads:                      2
+daily_folders:                 False
+snow_name:                     snow
+em_name:                       em
+variables:                     thickness, snow_density, specific_mass, liquid_water, temp_surf, temp_lower, temp_snowcover, thickness_lower, water_saturation, net_rad, sensible_heat, latent_heat, snow_soil, precip_advected, sum_eb, evaporation, snowmelt, swi, cold_content
+
+
+################################################################################
+# Parameters for restarting from crash
+################################################################################
+
+[isnobal restart]
+restart_crash:                 False
+depth_thresh:                  0.05
+wyh_restart_output:            None
+output_folders:                standard
+
+
+################################################################################
+# Running Python wrapped iSnobal
+################################################################################
+
+[ipysnobal]
+forcing_data_type:             netcdf
+
+
+################################################################################
+# Input constants for PySnobal
+################################################################################
+
+[ipysnobal constants]
+z_u:                           5.0
+z_t:                           5.0
+z_g:                           0.5
