+++ tests/basins/Lakes/gold_hrrr/make_gold_hrrr
@@ -1,25 +1,26 @@
 #!/usr/bin/env bash
 
-# A script to create the gold files needed for SMRF
+# A script to create the gold files needed for AWSM
 # Run this script in the main smrf directory, the config
 # file will put the outputs in the correct place
 #
 # Run as:
-#     $ ./tests/Lakes/gold_hrrr/make_gold_hrrr
+#     $ ./tests/basins/Lakes/gold_hrrr/make_gold_hrrr
 
 read -r -p "Recreating Lakes HRRR gold standard, are you sure? [y/N] " response
 if [[ "$response" = "yes" || "$response" = "y" ]]
 then
     echo "Updating Lakes HRRR gold files"
     echo "Removing old gold files"
-    rm tests/Lakes/gold_hrrr/*.nc
+    rm tests/basins/Lakes/gold_hrrr/*.nc
     
-    echo "Running SMRF"
-    run_smrf tests/Lakes/gold_hrrr/gold_config.ini
+    echo "Running AWSM"
+    awsm tests/basins/Lakes/gold_hrrr/gold_config.ini
 
     echo "Cleaning up"
-    rm tests/Lakes/gold_hrrr/log.txt
-    rm tests/Lakes/gold_hrrr/config.ini
+    mv tests/basins/Lakes/gold_hrrr/lakes/devel/wy2020/lakes_gold/runs/run0015_0017/*.nc \
+        tests/basins/Lakes/gold_hrrr
+    rm -rf tests/basins/Lakes/gold_hrrr/lakes
 else
     echo "Lakes HRRR gold files not updated"
 fi
